Interim commit, prod code compile and unit tests pass.
TBD: Fix production API Check integration tests via mock only make pub mock
feature blocks.
